+++ custom_components/ns_reisadvies/www/ns-reisadvies-card.js
@@ -2435,7 +2435,7 @@ window.customCards.push({
 });
 
 console.info(
-  "%c NS-REISADVIES-CARD %c v2.16.5 ",
+  "%c NS-REISADVIES-CARD %c v2.16.6 ",
   "color: white; background: #003082; font-weight: 700;",
   "color: #003082; background: #FFC917; font-weight: 700;"
 );